Getting NVFORTRAN compile was much more problematic

+++ src/commonmodule.f90
@@ -5,7 +5,7 @@ module CommonModule
   use, intrinsic :: iso_fortran_env, only: int32, int64
   implicit none
   public
-  integer, parameter  :: K_LEAF_CAPACITY = 16 !> 32 is better than 64  
+  integer(kind=int64), parameter  :: K_LEAF_CAPACITY = 16 !> 32 is better than 64  
   integer,  parameter :: K_COORDINATE_KIND = int32
 end module CommonModule
